No longer tracking Shader state
It greatly reduces how ergonomic the API is. It is better left to the dependees
of this library.

Tracking state through the type system is inflexible because it requires values
to be owned by value. Often people will want to pass them by (mutable)
references instead.

We don't know what state is important. We don't know how the dependee wants to
store the state. Perhaps it makes sense to compress multiple aspects of the
state into a single field. Like storing the compile status of multiple programs
in a bit-vector.